Wire synchs over libmempool-rs (Narwhal-style payload refs)
`types::Block` now references a single `BatchHash<Transaction>` in
`Body` rather than inlining `Vec<Hash<Transaction>>`. The canonical
payload lives in the referenced `Batch`, persisted in
`libstorage::Store` by libmempool-rs's `Processor`. Proposals carry
the batch inline (following leto-rs) so followers don't need a sync
round-trip on the happy path; relays/votes are hash-only.

Touches types, config-less; synchs is the pilot:

  - `Storage` loses its mempool surface (`pending_tx`, `add_transaction`,
    `cleave`, `clear`, `get_tx_pool_size`, `feature = "mempool"`).
    Transactions now flow client -> Mempool -> Batcher -> Processor
    -> `BatchHash` -> consensus, bypassing `Storage` entirely.
  - `Round` / `Height` / `View` change from `usize` to `u64`.
    `libmempool::Round` requires `Sub<Output=Self>` and `const MIN`
    through a macro that covers `u8..u128`/`i8..i128` -- not `usize`.
    Fixed-width also gives us a stable wire size across 32/64-bit.
  - `types::CountSealer` replaces the byte-based `libmempool::Sized`
    sealer so the historical `config.block_size` (txs per block) maps
    directly to batch size.
  - `ClientMsg::NewBlock(Block, Vec<Hash<Transaction>>, Payload)`
    carries tx hashes hydrated on the node side at commit time, so
    existing latency-tracking client logic survives unchanged.
  - Per-node batch store is a rocksdb at `<config-dir>/node-<id>.rocksdb`
    (override via `--store`).

Client tx submission flips from `TlsReliableSender<Replica, Transaction>`
to `TcpSimpleSender<Replica, Transaction>` because libmempool-rs's
client listener uses `TcpReceiver`. The node->client commit-notification
path stays on TLS.

Temporarily excludes apollo/artemis/optsync from the workspace + drops
the stress-test's preflight binary check for them; will be ported in
follow-up commits.

Synchs pilot results (50k tx, block_size=400, payload=0):
  - n=3, f=1: 57370 tx/s, 157 ms latency
  - n=7, f=3: 25244 tx/s, 381 ms latency
Pre-libmempool baseline on the same branch was 18k / 11k; the new
Narwhal flow removes leader-side cleave() and reactor-side
add_transaction churn and shrinks proposal size.

+++ types/src/msg/apollo/mod.rs
@@ -16,4 +16,5 @@ pub type Transaction = super::Transaction;
 pub type Replica = crate::Replica;
 pub type Storage = super::Storage<Block, Transaction>;
 pub type Payload = super::Payload;
-pub type Round = usize;
+/// See `types::Height` for why this is `u64` rather than `usize`.
+pub type Round = u64;
